+++ packages/web3-eth-contract/CHANGELOG.md
@@ -190,3 +190,4 @@ const transactionHash = receipt.transactionHash;
 
 -   Emit past contract events based on `fromBlock` when passed to `contract.events.someEventName` (#5201)
 -   Use different types for `ContractOptions` -> `jsonInterface` setter and getter (#5474)
+-   An issue within the `Contract` constructor where `provider` wasn't being set when provided within the `optionsOrContextOrReturnFormat` argument (#5669)